unit tests: Enables a few kubelet unit tests for Windows
Some of the kubelet tests are currently running only for Linux. We can
enable them for Windows as well.
